-I_StartSound should take a float for volume, not an int. -This fixes a recent problem where player originating sounds like grunts and weapon sounds would be muted. -Thanks AltDeath and 40oz for your troubleshooting.
SVN r2349 (0.6-dev)

+++ client/sdl/i_sound.cpp
@@ -271,7 +271,7 @@ void I_SetSfxVolume (float volume)
 // Pitching (that is, increased speed of playback)
 //  is set, but currently not used by mixing.
 //
-int I_StartSound (int id, int vol, int sep, int pitch, bool loop)
+int I_StartSound (int id, float vol, int sep, int pitch, bool loop)
 {
 	if(!sound_initialized)
 		return 0;
